+++ src/intelliSense/configDiagnosticProvider.ts
@@ -3,7 +3,7 @@
 
 "use strict";
 import * as fse from "fs-extra";
-import * as parser from "jsonc-parser/lib/umd/main";
+import * as parser from "jsonc-parser";
 import * as path from "path";
 import * as vscode from "vscode";
 import { Constants } from "../common/constants";
@@ -12,9 +12,10 @@ import { IntelliSenseUtility } from "./intelliSenseUtility";
 
 export class ConfigDiagnosticProvider {
     public async updateDiagnostics(document: vscode.TextDocument, diagCollection: vscode.DiagnosticCollection) {
-        if (!document && path.basename(document.uri.fsPath) !== Constants.deploymentTemplate
+        if (!document ||
+            (path.basename(document.uri.fsPath) !== Constants.deploymentTemplate
             && path.basename(document.uri.fsPath) !== Constants.deploymentDebugTemplate
-            && path.basename(document.uri.fsPath) !== Constants.moduleManifest) {
+            && path.basename(document.uri.fsPath) !== Constants.moduleManifest)) {
             return;
         }